## Summary

I implement the option proposed by Raph with a new RopeSlice struct to improve memory efficiency of rope diffing.

I keep an eye on performance and it's may be not what you want as you can see bellow.

**Benchmark before my commit:**
running 10 tests
test hash_diff      ... bench:       1,266 ns/iter (+/- 463)
test hash_diff_big  ... bench:     183,348 ns/iter (+/- 30,961)
test hash_diff_med  ... bench:      29,436 ns/iter (+/- 8,489)
test ne_idx_avx     ... bench:       5,297 ns/iter (+/- 834)
test ne_idx_detect  ... bench:       5,258 ns/iter (+/- 682)
test ne_idx_rev_sse ... bench:      13,240 ns/iter (+/- 2,105)
test ne_idx_rev_sw  ... bench:      68,948 ns/iter (+/- 6,669)
test ne_idx_sse     ... bench:       9,636 ns/iter (+/- 1,413)
test ne_idx_sw      ... bench:           0 ns/iter (+/- 0)
test scanner        ... bench:      11,012 ns/iter (+/- 4,030)

**Benchmark after my commit:**
running 10 tests
test hash_diff      ... bench:       1,127 ns/iter (+/- 332)
test hash_diff_big  ... bench:   6,765,850 ns/iter (+/- 317,297)
test hash_diff_med  ... bench:     192,641 ns/iter (+/- 44,295)
test ne_idx_avx     ... bench:       5,111 ns/iter (+/- 468)
test ne_idx_detect  ... bench:       4,988 ns/iter (+/- 477)
test ne_idx_rev_sse ... bench:      12,778 ns/iter (+/- 1,155)
test ne_idx_rev_sw  ... bench:      68,279 ns/iter (+/- 6,189)
test ne_idx_sse     ... bench:       6,288 ns/iter (+/- 3,110)
test ne_idx_sw      ... bench:           0 ns/iter (+/- 0)
test scanner        ... bench:      11,032 ns/iter (+/- 3,316)

The performance drop come from removing whitespace from my slice **(diff.rs line 91)** which make me create a new RopeSlice (or make my previous slice mutable).

I didn't manage to remove the overhead but I'm totally new to rust so I hope you have idea to solve it or I can try the second option proposed to improve the memory efficiency of rope diffing.

## Related Issues
`closes #946 `

## Review Checklist
- [ ] I have responded to reviews and made changes where appropriate.
- [x] I have tested the code with `cargo test --all`.
- [x] I have updated comments / documentation related to the changes I made.
- [x] I have rebased my PR branch onto xi-editor/master.