@cmyr 2019-03-19
Thanks for this! will take a look soon and see if we can figure out how to address the performance questions.