@HoNile 2019-03-19
Humm ... I'm still trying to improve this but look like I commit too early not sure it's worth it for you to look at this. Probably way more than performance problem.