@HoNile 2019-03-19
Now look better I hope I didn't miss something else.

running 10 tests
test hash_diff      ... bench:       1,197 ns/iter (+/- 149)
test hash_diff_big  ... bench:     313,091 ns/iter (+/- 47,905)
test hash_diff_med  ... bench:      56,170 ns/iter (+/- 6,490)
test ne_idx_avx     ... bench:       6,162 ns/iter (+/- 706)
test ne_idx_detect  ... bench:       5,764 ns/iter (+/- 1,112)
test ne_idx_rev_sse ... bench:      16,766 ns/iter (+/- 3,167)
test ne_idx_rev_sw  ... bench:      68,587 ns/iter (+/- 7,985)
test ne_idx_sse     ... bench:      10,005 ns/iter (+/- 1,663)
test ne_idx_sw      ... bench:           0 ns/iter (+/- 0)
test scanner        ... bench:      12,574 ns/iter (+/- 2,174)