@HoNile 2019-03-21
I think I could improve syntax with `while let` or `if let` at some point and non_ws_offset was better before my last change, but I don't want to change anything before review.