@HoNile 2019-03-24
I'm trying to do this but it look hard to return a `HashMap<Cow<str>, usize>` without allocation or without lifetime issue, as I can't insert line from `rope.lines_raw(..)` directly into my HashMap (need to insert something based on line[non_ws..]).