@HoNile 2019-03-26
Okay sorry I didn't relize that we can split COW in two categories during this work, my own version did `Cow::from(s[non_ws..].to_owned())` for all lines.
I update make_line_hashes with your version and just add comment and a parameter to not process all the rope, based on preliminary scan like for target.